pyproject.toml
[project]
name = "network-debugging"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "paramiko>=4.0.0",
    "python-dotenv>=1.0.0",
]
